refactor(ale_claw): fold computer_handler into tools/, drop computer/
computer_handler (OpenClawComputerHandler + MCPComputerHandler) is the handler
for the `computer` tool (assembled in tools/tools.py) and depends on
tools/mcp_runtime. Co-locate it with the substrate backends (fs_backends,
mcp_runtime) in tools/ rather than in a 1-file computer/ package that didn't
own the MCP transport/bridges anyway. Behavior-preserving: git mv + import
updates (..tools.mcp_runtime -> sibling .mcp_runtime; importers -> .tools.*).
928 passed / 3 skipped, VM smoke 1.0. README updated.
